+++ test/libsolidity/syntaxTests/functionTypes/call_value_on_non_payable_function_type.sol
@@ -1,7 +1,7 @@
 contract C {
     function (uint) external returns (uint) x;
     function f() public {
-        x.value(2)();
+        x.value(2)(1);
     }
 }
 // ----